Migrate cluster init to the behave fixture
Before this patch we run all tests against standard demo cluster. But
some tests (e.g. gpexpand or gpaddmirrors) don't use precreated cluster
at all. Tests tagged with @concourse_cluster always require custom
multihost configuration. So, let's refactor it now.

+++ gpMgmt/test/behave_utils/ci/fixtures.py
@@ -2,10 +2,13 @@ from behave import fixture
 
 
 @fixture
-def init_cluster(context):
-    context.execute_steps(u"""
-    Given the database is not running
-        And a working directory of the test as '/tmp/concourse_cluster'
-        And the user runs command "rm -rf ~/gpAdminLogs/gpinitsystem*"
-        And a cluster is created with mirrors on "cdw" and "sdw1, sdw2, sdw3"
-    """)
+def init_cluster(context, default=False):
+    if default:
+        context.execute_steps("Given a standard local demo cluster is created")
+    else:
+        context.execute_steps(u"""
+        Given the database is not running
+            And a working directory of the test as '/tmp/concourse_cluster'
+            And the user runs command "rm -rf ~/gpAdminLogs/gpinitsystem*"
+            And a cluster is created with mirrors on "cdw" and "sdw1, sdw2, sdw3"
+        """)
